requirements.txt
torch>=2.5
torchvision>=0.20

numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
scipy>=1.10
matplotlib>=3.7
seaborn>=0.13
PyYAML>=6.0
opencv-python>=4.8
pycocotools>=2.0
onnx>=1.14
onnxruntime>=1.16
timm>=0.9
faiss-cpu>=1.7
gensim>=4.3
optuna>=3.5

streamlit>=1.30
altair>=5.0
streamlit-agraph>=0.0.45

py2puml>=0.9

huggingface_hub>=0.20
kaggle>=1.6
kagglehub>=0.3

ftfy>=6.1
regex>=2023.0.0
tqdm>=4.65

git+https://github.com/openai/CLIP.git
git+https://github.com/facebookresearch/sam2.git